chore: creates MainPageBanner component
